:shirt: Rename to simpler names
